refactor: create `reth-exex`
