+++ crates/exex/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "reth-exex"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+description = "Execution extensions for Reth"
+
+[lints]
+workspace = true
+
+[dependencies]
+reth-config.workspace = true
+reth-node-api.workspace = true
+reth-node-core.workspace = true
+reth-primitives.workspace = true
+reth-provider.workspace = true
+reth-tasks.workspace = true
+tokio.workspace = true
+tokio-util.workspace = true
