+++ crates/node-builder/Cargo.toml
@@ -16,6 +16,7 @@ workspace = true
 reth-auto-seal-consensus.workspace = true
 reth-beacon-consensus.workspace = true
 reth-blockchain-tree.workspace = true
+reth-exex.workspace = true
 reth-provider.workspace = true
 reth-revm.workspace = true
 reth-db.workspace = true
