+++ crates/node-builder/src/builder.rs
@@ -7,7 +7,7 @@ use crate::{
         ComponentsBuilder, FullNodeComponents, FullNodeComponentsAdapter, NodeComponents,
         NodeComponentsBuilder, PoolBuilder,
     },
-    exex::{BoxedLaunchExEx, ExExContext},
+    exex::BoxedLaunchExEx,
     hooks::NodeHooks,
     node::FullNode,
     rpc::{RethRpcServerHandles, RpcContext, RpcHooks},
@@ -28,6 +28,7 @@ use reth_db::{
     test_utils::{create_test_rw_db, TempDatabase},
     DatabaseEnv,
 };
+use reth_exex::ExExContext;
 use reth_interfaces::p2p::either::EitherDownloader;
 use reth_network::{NetworkBuilder, NetworkConfig, NetworkEvents, NetworkHandle};
 use reth_node_api::{FullNodeTypes, FullNodeTypesAdapter, NodeTypes};
