+++ crates/node-builder/src/exex.rs
@@ -30,43 +30,8 @@
 
 use crate::FullNodeTypes;
 use futures::{future::BoxFuture, FutureExt};
-use reth_node_core::{
-    dirs::{ChainPath, DataDirPath},
-    node_config::NodeConfig,
-};
-use reth_primitives::{BlockNumber, Head};
-use reth_tasks::TaskExecutor;
+use reth_exex::ExExContext;
 use std::future::Future;
-
-/// Events emitted by an ExEx.
-#[derive(Debug)]
-pub enum ExExEvent {
-    /// Highest block processed by the ExEx.
-    ///
-    /// The ExEx must guarantee that it will not require all earlier blocks in the future, meaning
-    /// that Reth is allowed to prune them.
-    ///
-    /// On reorgs, it's possible for the height to go down.
-    FinishedHeight(BlockNumber),
-}
-
-/// Captures the context that an ExEx has access to.
-#[derive(Clone, Debug)]
-pub struct ExExContext<Node: FullNodeTypes> {
-    /// The current head of the blockchain at launch.
-    pub head: Head,
-    /// The configured provider to interact with the blockchain.
-    pub provider: Node::Provider,
-    /// The task executor of the node.
-    pub task_executor: TaskExecutor,
-    /// The data dir of the node.
-    pub data_dir: ChainPath<DataDirPath>,
-    /// The config of the node
-    pub config: NodeConfig,
-    /// The loaded node config
-    pub reth_config: reth_config::Config,
-    // TODO(alexey): add pool, payload builder, anything else?
-}
 
 /// A trait for launching an ExEx.
 trait LaunchExEx<Node: FullNodeTypes>: Send {
